Use shared LVGL stub for host tests
