+++ components/lvgl/lvgl.c
@@ -46,6 +46,17 @@ void lv_label_set_text_fmt(lv_obj_t *label, const char *fmt, ...) {
     va_end(args);
 }
 
+void lv_obj_add_event_cb(lv_obj_t *obj, lv_event_cb_t cb) {
+    if (obj) obj->event_cb = cb;
+}
+
+void lv_event_send(lv_obj_t *obj) {
+    if (obj && obj->event_cb) {
+        lv_event_t e = { .target = obj };
+        obj->event_cb(&e);
+    }
+}
+
 /* Additional stubs for the example project */
 
 void lv_init(void) {
